requirements.txt
# Core Airflow + dependencies
apache-airflow==2.10.2

# Data manipulation
pandas>=2.2.0

# Google Cloud clients
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.14.0

# DBT
dbt-core>=1.8.0
dbt-bigquery>=1.8.0

# Optional utilities
pyarrow>=15.0.0
